symmetry belongs to the lattice module
